♻️ refactor(RT buffer): Change namespace and remove const from readRT function
