+++ littlebot_base/include/littlebot_base/i_rt_buffer.hpp
@@ -25,7 +25,7 @@ public:
   virtual ~IRTBuffer() = default;
 
   // RT context
-  virtual const T * readRT() const noexcept = 0;
+  virtual const T * readRT() noexcept = 0;
 
   // Non-RT context
   virtual void writeNonRT(const T & data) = 0;
